[cleanup] Better Android auto manifest
